GH-10: Move data fetcher to separate file.
